rename hash.js to getHash.js
