+++ .eslintrc
@@ -21,7 +21,8 @@
     "import/no-extraneous-dependencies": [0, { "devDependencies": true }],
     "import/prefer-default-export": 0,
     "no-param-reassign": 0,
-    "class-methods-use-this": 0
+    "class-methods-use-this": 0,
+    "no-console": 0
   },
   "settings": {
     "react": {
