+++ src/config/getHash.js
